django-admin/requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
cloudflare==2.9.12
Django==4.1
django-admin-interface==0.20.0
django-appconf==1.0.5
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-imagekit==4.1.0
django-material==1.11.3
django-parler==2.3
django-rest-framework==0.1.0
django-rosetta==0.9.8
django-tinymce==3.5.0
djangorestframework==3.13.1
idna==3.3
jsonlines==3.1.0
pilkit==2.0
Pillow==10.4.0
polib==1.1.1
psycopg2-binary==2.9.9
pytz==2022.2.1
PyYAML==6.0.2
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.12
gunicorn==20.1.0
